+++ owncloudApp/src/main/java/com/owncloud/android/presentation/settings/automaticuploads/SettingsPictureUploadsFragment.kt
@@ -3,8 +3,9 @@
  *
  * @author Juan Carlos Garrote Gascón
  * @author Aitor Ballesteros Pavón
+ * @author Jorge Aguado Recio
  *
- * Copyright (C) 2023 ownCloud GmbH.
+ * Copyright (C) 2024 ownCloud GmbH.
  * <p>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -48,7 +49,10 @@ import com.owncloud.android.db.PreferenceManager.PREF__CAMERA_PICTURE_UPLOADS_PA
 import com.owncloud.android.db.PreferenceManager.PREF__CAMERA_PICTURE_UPLOADS_SOURCE
 import com.owncloud.android.db.PreferenceManager.PREF__CAMERA_PICTURE_UPLOADS_WIFI_ONLY
 import com.owncloud.android.domain.automaticuploads.model.UploadBehavior
+import com.owncloud.android.extensions.collectLatestLifecycleFlow
 import com.owncloud.android.extensions.showAlertDialog
+import com.owncloud.android.extensions.showMessageInSnackbar
+import com.owncloud.android.presentation.accounts.ManageAccountsViewModel
 import com.owncloud.android.ui.activity.FolderPickerActivity
 import com.owncloud.android.utils.DisplayUtils
 import kotlinx.coroutines.launch
@@ -59,6 +63,7 @@ class SettingsPictureUploadsFragment : PreferenceFragmentCompat() {
 
     // ViewModel
     private val picturesViewModel by viewModel<SettingsPictureUploadsViewModel>()
+    private val manageAccountsViewModel by viewModel<ManageAccountsViewModel>()
 
     private var prefEnablePictureUploads: SwitchPreferenceCompat? = null
     private var prefPictureUploadsPath: Preference? = null
@@ -69,6 +74,7 @@ class SettingsPictureUploadsFragment : PreferenceFragmentCompat() {
     private var prefPictureUploadsAccount: ListPreference? = null
     private var prefPictureUploadsLastSync: Preference? = null
     private var spaceId: String? = null
+    private lateinit var selectedAccount: String
 
     private val selectPictureUploadsPathLauncher =
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
@@ -103,10 +109,7 @@ class SettingsPictureUploadsFragment : PreferenceFragmentCompat() {
             ).toTypedArray()
             entryValues = listOf(UploadBehavior.COPY.name, UploadBehavior.MOVE.name).toTypedArray()
         }
-        prefPictureUploadsAccount = findPreference<ListPreference>(PREF__CAMERA_PICTURE_UPLOADS_ACCOUNT_NAME)?.apply {
-            entries = picturesViewModel.getLoggedAccountNames()
-            entryValues = picturesViewModel.getLoggedAccountNames()
-        }
+        prefPictureUploadsAccount = findPreference(PREF__CAMERA_PICTURE_UPLOADS_ACCOUNT_NAME)
 
         val comment = getString(R.string.prefs_camera_upload_source_path_title_required)
         prefPictureUploadsSourcePath?.title = String.format(prefPictureUploadsSourcePath?.title.toString(), comment)
@@ -123,18 +126,40 @@ class SettingsPictureUploadsFragment : PreferenceFragmentCompat() {
     private fun initStateObservers() {
         viewLifecycleOwner.lifecycleScope.launch {
             viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
-                picturesViewModel.pictureUploads.collect { pictureUploadsConfiguration ->
-                    enablePictureUploads(pictureUploadsConfiguration != null)
-                    pictureUploadsConfiguration?.let {
-                        prefPictureUploadsAccount?.value = it.accountName
-                        prefPictureUploadsPath?.summary = picturesViewModel.getUploadPathString()
-                        prefPictureUploadsSourcePath?.summary = DisplayUtils.getPathWithoutLastSlash(it.sourcePath.toUri().path)
-                        prefPictureUploadsOnWifi?.isChecked = it.wifiOnly
-                        prefPictureUploadsOnCharging?.isChecked = it.chargingOnly
-                        prefPictureUploadsBehaviour?.value = it.behavior.name
-                        prefPictureUploadsLastSync?.summary = DisplayUtils.unixTimeToHumanReadable(it.lastSyncTimestamp)
-                        spaceId = it.spaceId
-                    } ?: resetFields()
+                collectLatestLifecycleFlow(manageAccountsViewModel.userQuotas) { listUserQuotas ->
+                    val availableAccounts = listUserQuotas.filter { it.available != -4L }
+                    prefPictureUploadsAccount?.apply {
+                        entries = availableAccounts.map { it.accountName }.toTypedArray()
+                        entryValues = availableAccounts.map { it.accountName }.toTypedArray()
+                    }
+
+                    if (availableAccounts.isEmpty()) {
+                        enablePictureUploads(false, true)
+                        showMessageInSnackbar(getString(R.string.prefs_automatic_uploads_not_available_users_light))
+                    } else {
+                        val currentAccount = manageAccountsViewModel.getCurrentAccount()?.name
+                        currentAccount?.let {
+                            selectedAccount = if (manageAccountsViewModel.checkUserLight(currentAccount)) {
+                                availableAccounts.first().accountName
+                            } else {
+                                currentAccount
+                            }
+                        }
+
+                        picturesViewModel.pictureUploads.collect { pictureUploadsConfiguration ->
+                            enablePictureUploads(pictureUploadsConfiguration != null, false)
+                            pictureUploadsConfiguration?.let {
+                                prefPictureUploadsAccount?.value = it.accountName
+                                prefPictureUploadsPath?.summary = picturesViewModel.getUploadPathString()
+                                prefPictureUploadsSourcePath?.summary = DisplayUtils.getPathWithoutLastSlash(it.sourcePath.toUri().path)
+                                prefPictureUploadsOnWifi?.isChecked = it.wifiOnly
+                                prefPictureUploadsOnCharging?.isChecked = it.chargingOnly
+                                prefPictureUploadsBehaviour?.value = it.behavior.name
+                                prefPictureUploadsLastSync?.summary = DisplayUtils.unixTimeToHumanReadable(it.lastSyncTimestamp)
+                                spaceId = it.spaceId
+                            } ?: resetFields()
+                        }
+                    }
                 }
             }
         }
@@ -145,7 +170,7 @@ class SettingsPictureUploadsFragment : PreferenceFragmentCompat() {
             val value = newValue as Boolean
 
             if (value) {
-                picturesViewModel.enablePictureUploads()
+                picturesViewModel.enablePictureUploads(selectedAccount)
                 showAlertDialog(
                     title = getString(R.string.common_important),
                     message = getString(R.string.proper_pics_folder_warning_camera_upload)
@@ -228,8 +253,11 @@ class SettingsPictureUploadsFragment : PreferenceFragmentCompat() {
         super.onDestroy()
     }
 
-    private fun enablePictureUploads(value: Boolean) {
+    private fun enablePictureUploads(value: Boolean, isLightUser: Boolean) {
         prefEnablePictureUploads?.isChecked = value
+        if (isLightUser) {
+            prefEnablePictureUploads?.isEnabled = false
+        }
         prefPictureUploadsPath?.isEnabled = value
         prefPictureUploadsOnWifi?.isEnabled = value
         prefPictureUploadsOnCharging?.isEnabled = value
@@ -248,4 +276,5 @@ class SettingsPictureUploadsFragment : PreferenceFragmentCompat() {
         prefPictureUploadsBehaviour?.value = UploadBehavior.COPY.name
         prefPictureUploadsLastSync?.summary = null
     }
+
 }