+++ owncloudApp/src/main/java/com/owncloud/android/presentation/settings/automaticuploads/SettingsVideoUploadsFragment.kt
@@ -3,8 +3,9 @@
  *
  * @author Juan Carlos Garrote Gascón
  * @author Aitor Ballesteros Pavón
+ * @author Jorge Aguado Recio
  *
- * Copyright (C) 2023 ownCloud GmbH.
+ * Copyright (C) 2024 ownCloud GmbH.
  * <p>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -48,7 +49,10 @@ import com.owncloud.android.db.PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_PATH
 import com.owncloud.android.db.PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_SOURCE
 import com.owncloud.android.db.PreferenceManager.PREF__CAMERA_VIDEO_UPLOADS_WIFI_ONLY
 import com.owncloud.android.domain.automaticuploads.model.UploadBehavior
+import com.owncloud.android.extensions.collectLatestLifecycleFlow
 import com.owncloud.android.extensions.showAlertDialog
+import com.owncloud.android.extensions.showMessageInSnackbar
+import com.owncloud.android.presentation.accounts.ManageAccountsViewModel
 import com.owncloud.android.ui.activity.FolderPickerActivity
 import com.owncloud.android.utils.DisplayUtils
 import kotlinx.coroutines.launch
@@ -59,6 +63,7 @@ class SettingsVideoUploadsFragment : PreferenceFragmentCompat() {
 
     // ViewModel
     private val videosViewModel by viewModel<SettingsVideoUploadsViewModel>()
+    private val manageAccountsViewModel by viewModel<ManageAccountsViewModel>()
 
     private var prefEnableVideoUploads: SwitchPreferenceCompat? = null
     private var prefVideoUploadsPath: Preference? = null
@@ -69,6 +74,7 @@ class SettingsVideoUploadsFragment : PreferenceFragmentCompat() {
     private var prefVideoUploadsAccount: ListPreference? = null
     private var prefVideoUploadsLastSync: Preference? = null
     private var spaceId: String? = null
+    private lateinit var selectedAccount: String
 
     private val selectVideoUploadsPathLauncher =
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
@@ -100,10 +106,7 @@ class SettingsVideoUploadsFragment : PreferenceFragmentCompat() {
             entries = listOf(getString(R.string.pref_behaviour_entries_keep_file), getString(R.string.pref_behaviour_entries_remove_original_file)).toTypedArray()
             entryValues = listOf(UploadBehavior.COPY.name, UploadBehavior.MOVE.name).toTypedArray()
         }
-        prefVideoUploadsAccount = findPreference<ListPreference>(PREF__CAMERA_VIDEO_UPLOADS_ACCOUNT_NAME)?.apply {
-            entries = videosViewModel.getLoggedAccountNames()
-            entryValues = videosViewModel.getLoggedAccountNames()
-        }
+        prefVideoUploadsAccount = findPreference<ListPreference>(PREF__CAMERA_VIDEO_UPLOADS_ACCOUNT_NAME)
 
         val comment = getString(R.string.prefs_camera_upload_source_path_title_required)
         prefVideoUploadsSourcePath?.title = String.format(prefVideoUploadsSourcePath?.title.toString(), comment)
@@ -120,18 +123,40 @@ class SettingsVideoUploadsFragment : PreferenceFragmentCompat() {
     private fun initLiveDataObservers() {
         viewLifecycleOwner.lifecycleScope.launch {
             viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
-                videosViewModel.videoUploads.collect { videoUploadsConfiguration ->
-                    enableVideoUploads(videoUploadsConfiguration != null)
-                    videoUploadsConfiguration?.let {
-                        prefVideoUploadsAccount?.value = it.accountName
-                        prefVideoUploadsPath?.summary = videosViewModel.getUploadPathString()
-                        prefVideoUploadsSourcePath?.summary = DisplayUtils.getPathWithoutLastSlash(it.sourcePath.toUri().path)
-                        prefVideoUploadsOnWifi?.isChecked = it.wifiOnly
-                        prefVideoUploadsOnCharging?.isChecked = it.chargingOnly
-                        prefVideoUploadsBehaviour?.value = it.behavior.name
-                        prefVideoUploadsLastSync?.summary = DisplayUtils.unixTimeToHumanReadable(it.lastSyncTimestamp)
-                        spaceId = it.spaceId
-                    } ?: resetFields()
+                collectLatestLifecycleFlow(manageAccountsViewModel.userQuotas) { listUserQuotas ->
+                    val availableAccounts = listUserQuotas.filter { it.available != -4L }
+                    prefVideoUploadsAccount?.apply {
+                        entries = availableAccounts.map { it.accountName }.toTypedArray()
+                        entryValues = availableAccounts.map { it.accountName }.toTypedArray()
+                    }
+
+                    if (availableAccounts.isEmpty()) {
+                        enableVideoUploads(false, true)
+                        showMessageInSnackbar(getString(R.string.prefs_automatic_uploads_not_available_users_light))
+                    } else {
+                        val currentAccount = manageAccountsViewModel.getCurrentAccount()?.name
+                        currentAccount?.let {
+                            selectedAccount = if (manageAccountsViewModel.checkUserLight(currentAccount)) {
+                                availableAccounts.first().accountName
+                            } else {
+                                currentAccount
+                            }
+                        }
+
+                        videosViewModel.videoUploads.collect { videoUploadsConfiguration ->
+                            enableVideoUploads(videoUploadsConfiguration != null, false)
+                            videoUploadsConfiguration?.let {
+                                prefVideoUploadsAccount?.value = it.accountName
+                                prefVideoUploadsPath?.summary = videosViewModel.getUploadPathString()
+                                prefVideoUploadsSourcePath?.summary = DisplayUtils.getPathWithoutLastSlash(it.sourcePath.toUri().path)
+                                prefVideoUploadsOnWifi?.isChecked = it.wifiOnly
+                                prefVideoUploadsOnCharging?.isChecked = it.chargingOnly
+                                prefVideoUploadsBehaviour?.value = it.behavior.name
+                                prefVideoUploadsLastSync?.summary = DisplayUtils.unixTimeToHumanReadable(it.lastSyncTimestamp)
+                                spaceId = it.spaceId
+                            } ?: resetFields()
+                        }
+                    }
                 }
             }
         }
@@ -142,7 +167,7 @@ class SettingsVideoUploadsFragment : PreferenceFragmentCompat() {
             val value = newValue as Boolean
 
             if (value) {
-                videosViewModel.enableVideoUploads()
+                videosViewModel.enableVideoUploads(selectedAccount)
                 showAlertDialog(
                     title = getString(R.string.common_important),
                     message = getString(R.string.proper_videos_folder_warning_camera_upload)
@@ -225,8 +250,11 @@ class SettingsVideoUploadsFragment : PreferenceFragmentCompat() {
         super.onDestroy()
     }
 
-    private fun enableVideoUploads(value: Boolean) {
+    private fun enableVideoUploads(value: Boolean, isLightUser: Boolean) {
         prefEnableVideoUploads?.isChecked = value
+        if (isLightUser) {
+            prefEnableVideoUploads?.isEnabled = false
+        }
         prefVideoUploadsPath?.isEnabled = value
         prefVideoUploadsOnWifi?.isEnabled = value
         prefVideoUploadsOnCharging?.isEnabled = value
@@ -245,4 +273,5 @@ class SettingsVideoUploadsFragment : PreferenceFragmentCompat() {
         prefVideoUploadsBehaviour?.value = UploadBehavior.COPY.name
         prefVideoUploadsLastSync?.summary = null
     }
+
 }