+++ owncloudApp/src/main/java/com/owncloud/android/ui/activity/DrawerActivity.kt
@@ -312,15 +312,21 @@ abstract class DrawerActivity : ToolbarActivity() {
                 is UIResult.Success -> {
                     uiResult.data?.let { userQuota ->
                         when {
+                            userQuota.available == -4L -> { // Light users (oCIS)
+                                getAccountQuotaText()?.text = getString(R.string.drawer_unavailable_used_storage)
+                                getAccountQuotaBar()?.isVisible = false
+                                getAccountQuotaStatusText()?.isVisible = false
+                            }
+
                             userQuota.available < 0 -> { // Pending, unknown or unlimited free storage
                                 getAccountQuotaBar()?.apply {
-                                    isVisible = true
-                                    progress = 0
-                                    progressTintList = ColorStateList.valueOf(resources.getColor(R.color.color_accent))
+                                   isVisible = true
+                                   progress = 0
+                                   progressTintList = ColorStateList.valueOf(resources.getColor(R.color.color_accent))
                                 }
                                 getAccountQuotaText()?.text = String.format(
-                                    getString(R.string.drawer_unavailable_free_storage),
-                                    DisplayUtils.bytesToHumanReadable(userQuota.used, this, true)
+                                   getString(R.string.drawer_unavailable_free_storage),
+                                   DisplayUtils.bytesToHumanReadable(userQuota.used, this, true)
                                 )
                                 getAccountQuotaStatusText()?.visibility = View.GONE
                             }