+++ owncloudApp/src/main/java/com/owncloud/android/ui/activity/ReceiveExternalFilesActivity.java
@@ -10,6 +10,7 @@
  * @author Abel García de Prada
  * @author John Kalimeris
  * @author Aitor Ballesteros Pavón
+ * @author Jorge Aguado Recio
  *
  * Copyright (C) 2012  Bartek Przybylski
  * Copyright (C) 2024 ownCloud GmbH.
@@ -72,7 +73,6 @@ import com.owncloud.android.data.providers.implementation.OCSharedPreferencesPro
 import com.owncloud.android.db.PreferenceManager;
 import com.owncloud.android.domain.exceptions.UnauthorizedException;
 import com.owncloud.android.domain.files.model.OCFile;
-import com.owncloud.android.domain.spaces.model.OCSpace;
 import com.owncloud.android.extensions.ActivityExtKt;
 import com.owncloud.android.extensions.ThrowableExtKt;
 import com.owncloud.android.lib.common.OwnCloudAccount;
@@ -117,6 +117,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import static com.owncloud.android.presentation.settings.advanced.SettingsAdvancedFragment.PREF_SHOW_HIDDEN_FILES;
+import static org.koin.core.parameter.ParametersHolderKt.parametersOf;
 import static org.koin.java.KoinJavaComponent.get;
 import static org.koin.java.KoinJavaComponent.inject;
 
@@ -179,7 +180,7 @@ public class ReceiveExternalFilesActivity extends FileActivity
 
     private boolean showHiddenFiles;
     private OCSharedPreferencesProvider sharedPreferencesProvider;
-    private OCSpace personalSpace;
+    private boolean areSpacesAllowed;
 
 
     Pattern pattern = Pattern.compile("[/\\\\]");
@@ -221,17 +222,15 @@ public class ReceiveExternalFilesActivity extends FileActivity
         }
         mSortOptionsView.setVisibility(View.GONE);
 
-        mReceiveExternalFilesViewModel = get(ReceiveExternalFilesViewModel.class);
-        subscribeToViewModels();
 
         initPickerListener();
 
     }
     private void subscribeToViewModels() {
-        mReceiveExternalFilesViewModel.getPersonalSpaceLiveData().observe(this, ocSpace -> {
-            personalSpace = ocSpace;
+        mReceiveExternalFilesViewModel.getSpacesAreAllowed().observe(this, spaces -> {
+            areSpacesAllowed = spaces;
 
-            if (personalSpace == null) { // OC10 Server
+            if (!areSpacesAllowed) { // OC10 Server
                 showListOfFiles();
                 showRetainerFragment();
                 updateDirectoryList();
@@ -337,7 +336,8 @@ public class ReceiveExternalFilesActivity extends FileActivity
     @Override
     protected void onAccountSet(boolean stateWasRecovered) {
         super.onAccountSet(mAccountWasRestored);
-        mReceiveExternalFilesViewModel.getPersonalSpaceForAccount(getAccount().name);
+        mReceiveExternalFilesViewModel = get(ReceiveExternalFilesViewModel.class, null , () -> parametersOf(getAccount().name));
+        subscribeToViewModels();
         initTargetFolder();
 
         mReceiveExternalFilesViewModel.getSyncFolderLiveData().observe(this, eventUiResult -> {
@@ -360,7 +360,7 @@ public class ReceiveExternalFilesActivity extends FileActivity
                 } else if (uiResult instanceof UIResult.Success) {
                     mSyncInProgress = false;
                     updateDirectoryList();
-                    if (mParents.size() == 1 && personalSpace == null) {
+                    if (mParents.size() == 1 && !areSpacesAllowed) {
                         updateToolbar(getString(R.string.uploader_top_message));
                     }
                     if(fragmentContainer.getVisibility() == View.VISIBLE) {
@@ -433,7 +433,6 @@ public class ReceiveExternalFilesActivity extends FileActivity
                     onAccountSet(mAccountWasRestored);
                     dialog.dismiss();
                     PreferenceManager.setLastUploadPath("/", this);
-                    mReceiveExternalFilesViewModel.getPersonalSpaceForAccount(getAccount().name);
                     mAccountSelected = true;
                     mAccountSelectionShowing = false;
                 });
@@ -458,7 +457,7 @@ public class ReceiveExternalFilesActivity extends FileActivity
             String full_path = generatePath(mParents);
             startSyncFolderOperation(getStorageManager().getFileByPath(full_path, currentSpaceId));
             updateDirectoryList();
-            if (mParents.size() <= 1 && personalSpace == null) {
+            if (mParents.size() <= 1 && !areSpacesAllowed) {
                 updateToolbar(getString(R.string.uploader_top_message));
             }
         }